requirements.txt
Flask==3.1.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
python-dotenv==1.1.1
Werkzeug==3.1.3
PyMySQL==1.1.1

sqlalchemy~=2.0.43